pppRyjMegaBirth: signed mesh-emit-index reset compare
birth 76.58->76.68 (+0.10). The meshEmitIndex>=pathInfo[1] reset guard:
keep pathInfo[1] signed via (int)pathInfo[1] and zero-extend
meshEmitIndex, so MWCC emits lha+cmpw (signed) matching the target
instead of clrlwi+cmplw (unsigned u16 mask).

+++ src/pppRyjMegaBirth.cpp
@@ -963,7 +963,7 @@ void birth(
 			if (pathBase != NULL) {
 				int sampleIndex;
 
-				if ((u16)work->m_meshEmitIndex >= (u16)pathInfo[1]) {
+				if ((int)(u16)work->m_meshEmitIndex >= (int)pathInfo[1]) {
 					work->m_meshEmitIndex = 0;
 				}
